requirements.txt
openai
openai-agents
python-dotenv
pandas
numpy
streamlit
altair
scikit-learn
mlxtend
openpyxl